CDAP-11341 Added ability to operate on string, so you can specify 'stemming <column>' on string column.
